Add alerts manager role and enforce permissions (#6215)
Closes #6017

+++ graylog2-server/src/main/java/org/graylog/events/rest/EventNotificationsResource.java
@@ -21,12 +21,14 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.graylog.events.audit.EventsAuditEventTypes;
 import org.graylog.events.notifications.DBNotificationService;
 import org.graylog.events.notifications.NotificationDto;
 import org.graylog.events.notifications.NotificationResourceHandler;
 import org.graylog2.alarmcallbacks.EmailAlarmCallback;
 import org.graylog2.audit.jersey.AuditEvent;
+import org.graylog2.database.PaginatedList;
 import org.graylog2.plugin.alarms.callbacks.AlarmCallback;
 import org.graylog2.plugin.configuration.ConfigurationRequest;
 import org.graylog2.plugin.rest.PluginRestResource;
@@ -35,6 +37,7 @@ import org.graylog2.search.SearchQuery;
 import org.graylog2.search.SearchQueryField;
 import org.graylog2.search.SearchQueryParser;
 import org.graylog2.shared.rest.resources.RestResource;
+import org.graylog2.shared.security.RestPermissions;
 
 import javax.inject.Inject;
 import javax.validation.constraints.NotBlank;
@@ -90,13 +93,17 @@ public class EventNotificationsResource extends RestResource implements PluginRe
                                                                 @ApiParam(name = "per_page") @QueryParam("per_page") @DefaultValue("50") int perPage,
                                                                 @ApiParam(name = "query") @QueryParam("query") @DefaultValue("") String query) {
         final SearchQuery searchQuery = searchQueryParser.parse(query);
-        return PaginatedResponse.create("notifications", dbNotificationService.getAllPaginated(searchQuery, "title", page, perPage), query);
+        final PaginatedList<NotificationDto> result = dbNotificationService.searchPaginated(searchQuery, notification -> {
+            return isPermitted(RestPermissions.EVENT_NOTIFICATIONS_READ, notification.id());
+        }, "title", page, perPage);
+        return PaginatedResponse.create("notifications", result, query);
     }
 
     @GET
     @Path("/{notificationId}")
     @ApiOperation("Get a notification")
     public NotificationDto get(@ApiParam(name = "notificationId") @PathParam("notificationId") @NotBlank String notificationId) {
+        checkPermission(RestPermissions.EVENT_NOTIFICATIONS_READ, notificationId);
         return dbNotificationService.get(notificationId)
                 .orElseThrow(() -> new NotFoundException("Notification " + notificationId + " doesn't exist"));
     }
@@ -104,6 +111,7 @@ public class EventNotificationsResource extends RestResource implements PluginRe
     @POST
     @ApiOperation("Create new notification definition")
     @AuditEvent(type = EventsAuditEventTypes.EVENT_NOTIFICATION_CREATE)
+    @RequiresPermissions(RestPermissions.EVENT_NOTIFICATIONS_CREATE)
     public NotificationDto create(NotificationDto dto) {
         return resourceHandler.create(dto);
     }
@@ -114,6 +122,7 @@ public class EventNotificationsResource extends RestResource implements PluginRe
     @AuditEvent(type = EventsAuditEventTypes.EVENT_NOTIFICATION_UPDATE)
     public NotificationDto update(@ApiParam(name = "notificationId") @PathParam("notificationId") @NotBlank String notificationId,
                                   NotificationDto dto) {
+        checkPermission(RestPermissions.EVENT_NOTIFICATIONS_EDIT, notificationId);
         dbNotificationService.get(notificationId)
                 .orElseThrow(() -> new NotFoundException("Notification " + notificationId + " doesn't exist"));
 
@@ -129,6 +138,7 @@ public class EventNotificationsResource extends RestResource implements PluginRe
     @ApiOperation("Delete a notification")
     @AuditEvent(type = EventsAuditEventTypes.EVENT_NOTIFICATION_DELETE)
     public void delete(@ApiParam(name = "notificationId") @PathParam("notificationId") @NotBlank String notificationId) {
+        checkPermission(RestPermissions.EVENT_NOTIFICATIONS_DELETE, notificationId);
         resourceHandler.delete(notificationId);
     }
